Add a wiki-dev.o.o server to test newer mediawiki
The wiki-dev.openstack.org server will eventually be used to test
newer versions of Mediawiki and extensions/skins. To accommodate
this, also parameterize server backups so that they don't conflict
with production (and are in fact disabled entirely for the dev
site). Further, switch the class default for cert/key/chain files
from an empty string to undef so that they will fall through to
undef properly in the mediawiki module.

Change-Id: I6505d3af87f670e71a440c76873c085d97e5b82f