Update zuul-env on job nodes
Improve the zuul-env handling in openstack_project::slave_common to
also update packages when the zuul-env-reqs.txt file changes. This
allows us to increase the version of Zuul installed in the
virtualenv over time on persistent job nodes, though at the expense
of needing to specify a version rather than just using the master
branch tip.

Specify Zuul 2.5.0 from PyPI in the virtualenv's requirements list
so that we will trigger an update, since jobs running on some
persistent nodes need a newer Zuul release than was initially
installed on them.

Change-Id: If07b31f3a735cf7bcf6bfb8591ed37577f5ae201