launch-node.py: More verbose logging
One problem with "shell script as python" is that there's no
equivalent of "-x" in shell, which makes it really hard to extract
what's being called and where output came from.

This adds a bit more verbose logging around the ssh calls to try and
help someone parsing the logs.

Change-Id: I85e2415b47e044cfa1c678fc7786b4891fa1f93e