upgrade Gerrit's javamelody plugin on review-dev.o.o
upgrade Gerrit's javamelody plugin[1] on review-dev.o.o
This brings in an upstream change[2] to update javamelody to version
1.60.0

NOTE: Approving this change will kill Gerrit due to a bug, the change
to fix[3] this bug is uncommited atm.  The way to avoid the bug is
to stop Gerrit, approve this change, allow puppet to propogate
the update then start Gerrit.

[1] http://tarballs.openstack.org/ci/gerrit/plugins/javamelody/javamelody-1de5d37.jar
[2] https://gerrit.googlesource.com/plugins/javamelody/+/1de5d373afb1b1edf0647c804a92dc1ce8e96ec2
[3] https://review.openstack.org/#/c/268689/

Change-Id: I33102ce76521c5ffc25f83d3d2022f87f9732f18