Fix logstash-filters test
In order to run the script directly in the test, the script needs to
have the executable bit set. Currently, the test is failing with a
"Permission denied" error when attempting to run the script.

We also need to run install_modules.sh in the logstash-filters test, so
move that invocation to the shared prep-apply.sh script.

Change-Id: Ic1ab6d04232abb3cf347f0d27afbf9172f62b0e4