Add email servers to firehose
A requirement for running lpmqtt is to have an imap server that is
subscribed to launchpad notifications to use as a source for the event
stream. However there is no preexisting imap server setup anywhere to
use for this. This commit deploys the necessary components for the imap
server. Setting up the launchpad user and the subscription will be a
manual process.

Change-Id: I063ac9da89d227ac59ca1cb8765e3615d7bfdbba