Sort the users install list chronologically
The openstack_project:users_install class has a list of accounts
which are granted root access on all servers. This list has
traditionally seen new additions appended, leaving it ordered by the
precedence by which each member was tasked with this responsibility.
Adjustment the current list so that recent additions retain this
chronological ordering.

Change-Id: I365369ef9ce80d2d24cd6aa786da7215b1204e26