Disable mqtt output plugin on logstash workers
Logstash is busted and the last time it updated elasticsearch was when
we attempted to turn this plugin on. Whose bright idea was it to use
that anyway. This commit turns off the configuration for the mqtt output
plugin and reverts us to just using the elasticsearch output for now. We
can investigate adding this again at some point in the future.

Change-Id: I88d5e507c8b32a6969a48d6bf6e6881985898387