Add a note to docs for websockets being disabled
This commit adds a note to the firehose docs to document that we've
temporarily disabled access to the websockets ports. This way users
will know that the websockets access isn't expected to work right now.
Once we have a solution in place to address the DOS when using
websockets we can remove this note when we open up the firewall again.

Change-Id: Ia6fb1ca0f80873e5547d2e72676b1b4ebeb01521