Lengthen bandersnatch timeout
The old 30 minute timeout was too conservative because it is possible
for packages to push a lot of large updates that require more than half
an hour to sync. Bump the timeout to four hours which should handle
these specific cases more gracefully while still timing out the run in a
reasonable amount of time if something has gone wrong.

Change-Id: I19e53c576579442ab358f8824d55bdbe322e913b