Allow zuul launcher and scheduler on the same host
Refactor to allow both launcher and scheduler to be setup on the same
server. Since both zuul_laucher and zuul_scheduler calls '::zuul' we
move that call up a level so that it won't conflict where a parent
module needs to setup a zuul_launcher and zuul_schedular on the same
node.

The use case for this change is to setup zuul-dev.o.o to run both
zuul schedular and launcher.

depends-on: Ief4ad30d5bb7a48b56b24b804643d3419fd1f110
Change-Id: I47533709a9ddd54347868020b4a905140870ddd9