gerritbot: Refactor gerritbot class to use multiple
For a while we've run into a problem where gerritbot has reached the
maximum channels that a single user is allowed to be in. A simple
solution is to simply add another user, however the existing gerritbot
class does not allow for this (it is a singleton). Thus, a refactor to
use Puppet's defined types is the parent change for this one.

Specifically, this commit modifies the Puppet manifest to use the new
invocation format for the gerritbot module.

Change-Id: I82a09f6a7f92501dd3818f939db637526c77753e
Depends-On: I5a8cd48c13a9a750d0b5a301a3c7202c97ca9f22