Add pip install requests[security] package
The requests package does not automatically include requests[security].
If the release[security] package is not installed you will get an
insecure platform warning.

/packages/urllib3/util/ssl_.py:122: InsecurePlatformWarning: A true
SSLContext object is not available. This prevents urllib3 from
configuring SSL appropriately and may cause certain SSL connections
to fail. You can upgrade to a newer version of Python to solve this.

Change-Id: Ib97ca490069755c532f9d34a231f3375cbf164b7