[WIP] Create docker01.openstack.org
As docker is more used my projects in openstack, it would be great is
we could share images. As such, create docker01.openstack.org to allow
up to publish images too.

This patch resolves a small bug which didn't allow static.o.o to use
snakeoil certs which was discovered in adapting SSL to this.

Change-Id: Ia5b502b500a6ad7e50259c5964f81e8369195f10
Signed-off-by: Paul Belanger <pabelanger@redhat.com>