Allow links in the message
The negative matching group that's supposed to prevent the closing li
tag from matching gets confused when there is a link in the text. It
only matches up to the beginning of the link, but then there is no li
closing tag after that, so hijinx ensue.

Just replace it with a .* since the regex is already anchored by the
closing li tag.

Change-Id: I030b2f00189e1ffed9981b1c3e4540306517e1bf