Define Gerrit's gitweb configs as a dict
Change Ie470a38b refactored how Gerrit's gitweb configurations
are set, this change passes the gitweb parameter as a dictionary
of configurations.

This change also fixes  gitweb configuration on review-dev.o.o

The cgit link on the change screen is currenty pointing to
something like this:
  https://git.openstack.org/cgit/gtest-org/test.git/commit/?id=ef9a5661607b5413ef2254b7bfbcbbcdaad57fca

Which is incorrect (*notice 'test.git' in url), gtest-org/test is the
Gerrit project.  Adding the revision parameter will change the
reference to:
  https://git.openstack.org/cgit/gtest-org/test/commit/?id=ef9a5661607b5413ef2254b7bfbcbbcdaad57fca

The 2nd url will work, while the 1st one does not.

TODO: keeping cgit, web_repo_url, and web_repo_url_encode
in init.pp to keep puppet happy (as a transition step).
Those params are not used and should be removed in a
follow up change.

Change-Id: I56e22bc337e6a17db716234841e0a7f961f2e81a
depends-on: Ie470a38bc14a2b37a43f45be274ed193c1c00a1a