Run timeout with --foreground
There is a bug/behaviour/issue where ansible will end up locked-up
running under timeout if ssh processes attempt to write to /dev/tty.
ssh will do this to express it's unrest at an unknown host key, or
throw up a password prompt, for example.  We are seeing shade return a
few hosts that aren't configured to ansible, and connection attempts
to them then proceed to lock-up ansible.

The argument is:

 --foreground : when not running timeout directly from a shell prompt,
   allow COMMAND to read from the TTY and get TTY signals; in this
   mode, children of COMMAND will not be timed out

The failure case is actually running this from an interactive prompt,
because this doesn't seem to affect cron.  I've had enough terminals
for one day, but this is what libvirt did [1] and seems to make
"run_all.sh" work from a prompt with both ansible 2.0.2.0 and 2.2

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1025269

Change-Id: I183e2cc3f849fc04cda2e9bf8aa6efc05b020bfb