Add bindep virtualenv to wheel-build slaves
I've noticed that wheel jobs are trying to build wheels for which we
don't have dev packages installed -- but these packages are listed in
the requirements project bindep.txt.

AFAICS, we're not using this at all.  I think what we need to do is
get the wheel jobs to run the usual install-distro-packages macro,
which should install the packages provided by the requirements
project.  That macro assumes bindep is in this virtualenv as described
by the element in the comment.  This should keep the virtualenv
up-to-date for the long-lived wheel slaves.

Change-Id: Ied1dc4936037b9e296017c8e3734250cca820a7b