Add n101.openstack.org server
Stand up a server we'll be using for nodepool (zuulv3) testing.
Currently we don't have the puppet bits created, I'll continue working
on this, but they will be added in a follow up patchset.

We'll be using this for the Atlanta PTG with the goal of moving it
into production when our feature/zuulv3 branch is merged into master
for nodepool.

Change-Id: I7a3edb871510d1000a9bbf06944551d4b272ac8d
Signed-off-by: Paul Belanger <pabelanger@redhat.com>