Add devstack platform tests to openstack-health
I always wondered why I couldn't make a dashboard for centos/fedora
health and today figured it out ... because these "platform" jobs are
non-voting and in the check queue, logstash.o.o skips sending the
message to the subunit workers to process their testrepository
artifacts.  Ergo no stats in openstack-health.

Add a special matcher for check queue jobs with "dsvm-platform-" in it
to match these jobs (apart from rpm distros, we've also had trusty in
here during transition periods to "keep the lights on", although that
was just kicked out of master because of libvirt updates).

This will allow us to use the usual openstack-health and related db
queries to monitor the health of these platforms.

Change-Id: I11a7cb0cb93392c40bc89854eb36d31e67351263