Ectomy some Jenkins out of the docs
There are a bunch of places where the narrative text say things like
"jenkins runs jobs". This hasn't been true for a while and it's getting
less true.

Left third-party and running-your-own alone because those are
instructions for other people to do things - and we are not yet at the
point where we are suggesting anyone do their things like we do our
things.

Also renamed the logstash client config file. It's the only change in
here that actually affects running code, but it stood out when making
the docs changes.

Change-Id: Iaf2128c3f953976180c71cb599fcbff7bc06c28a