Add puppetry for translate01 xenialing
This is part of the Zanata upgrade process. We will be deploying new
Zanata to translate01.openstack.org so add node def in site.pp to
support that. This keeps the old translate.o.o def too so that it is
left alone untouched.

Note that this also adds code to pass through wildlfy's install URL and
fixes the specification of the wildfly version number. I think the
version number var being wrong was not a problem for us because the
wildfly installation only runs if there is no installation dir in place.

Change-Id: I4b6ababcc34e2464cf6a824e92a1c146d553dcaf