Add #openstack-deployment channel to logging
As per TC requirements, all channels in the #openstack-* freenode
namespace should be logged via OpenStack Infrastructure.

There seems to be a general consensus (although not total agreement)
that a openstack-deployment IRC channel would be benefecial for
cross-project collaboration between various deployment tool teams.

The channel has already been created and is in limited initial use.
To meet the TC requirements, this review is submitted for logging
purposes.

Reference:
http://lists.openstack.org/pipermail/openstack-dev/2017-February/112960.html

Depends-On: Ia688174292383144da4885596d332c7800e8f353
Change-Id: Ia688174292383144da4885596d332c7800e8f353