[WIP] centos7: don't install EPEL and use centos openstack repo
EPEL is not really helping us test, because upstream RDO (and hence
centos openstack) are designed to run without it.  In fact it seems to
actively get in the way

Also, we should use the centos openstack packages.  The RDO packages
have been a point of failure before, at least the centos ones are
mirrored.

[1] https://review.openstack.org/#/c/438325/2

Change-Id: If6cf1bc300c6b59a8defb09fb3a3a1254e392a43