Add provider defaults to launch-node.py
As we use more providers for the control plane, it becomes a bit
annoying to have to list images, flavors, regions etc to remember what
goes where and of what type.

With this in mind, keep a list of providers and their defaults in the
launch-node script.  You can select one by setting the
PROVIDER_DEFAULTS environment flag.  Any more specific command-line
flags will override the provider defaults.

Note there is a slight change here in that "--cloud" is no longer one
of those double-dash-optional-but-actually-required arguments.  Since
we were defaulting to RAX stuff for a bunch of other flags (that
wouldn't have worked with other clouds) it didn't seem very
consistent.  So we just use our default provider (currently
"rax-prod") as the default for everything if nothing is specified.

However, as a handbrake on just blasting ahead if you type this
without thinking too much, I've added a quick prompt with some
overview info requiring you to press 'y'.

Depends-On: Id89bbf6306adebe96c1db417a4b8b5ae51023af0
Change-Id: I6f56bfc8fb91058b756bae2e0ed920ef36858292