Switch docs.o.o to worker MPM
This is a follow-on from I064b81076bcfcb200e0990627fc1aa0be8bbf058

We see the same issues with "scoreboard is full" and segfaults/dropped
connections on this host.  It's a semi-frequent issue in

logs.o.o seems to be working fine with this, so switch here too.  We
can reconsider in Xenial timeframe when we've got a later apache.

Change-Id: I05f1b1731cb6358e2e37574a90692ffe91f07ad3