Stop saying we announce signing key transitions
Our procedural documentation doesn't have instructions for
generating and sending cross-signed key transition announcements to
the mailing list, and we haven't actually done it yet (even though
we've completed two transitions at this point). Remove the claim in
our documentation that we do it for now; at some later date we can
always re-add it and document a procedure for generating/sending it
if we decide it's really needed after all.

Change-Id: I7702f1e6713fd6c62223d2d245013678586c46db