Start firehose doc section for client on port 1883
This commit starts a new section in the firehose docs for client
examples on how to print all messages from the firehose to STDOUT. All
the other examples in the doc either use ssl or websockets. It's good
to have an example of how to just connect to the firehose on vanilla
mqtt.

Change-Id: Iba9fb10e75eae10efcd59b70b6b83589a077610f