Move package resources to o_p::server
Once nodepool is building worker images with these packages already
installed, we don't need to manage them in the common
openstack_project::template class. This patch moves them over to
openstack_project::server so that it is obvious that puppet doesn't need
to manage them for both single-use workers and long-lived servers.
Eventually the openstack_project::template class should be empty and we
can remove it.

Depends-On: Ie1a0aba57390c9c0b269b4cbb076090ae1de73a9
Change-Id: I31295cdf12941e4adcf87a73418df1d17d9ec3d2