Move NTP management into o_p::server
Once nodepool is building images with NTP already installed and enabled,
we don't need to manage it in the common openstack_project::template
class. This patch moves them over to openstack_project::server so that
it is obvious that puppet doesn't need to manage them for both
single-use workers and long-lived servers. Eventually the
openstack_project::template class should be empty and we can remove it.

Depends-On: Iee6babc183dd12cc82ce76ddfde04f2d98ddc4d6
Change-Id: Ie808a5b62014716c8506506fd15f39dba06e76b6