Use caching reverse proxy for rdo repos
We can't sanely mirror the rdo packages beacuse they are constantly
updating. Instead put a caching reverse proxy in front of the repos in
each region on our existing mirror servers. This should allow us to get
region local data that is "hot" and hopefully improve reliability to the
rdo repos.

Change-Id: I6ab810f7b6788459da1cee00baf6b053f31843cf