Set Gerrit offline_reindex to true
When doing an upgrade from Gerrit 2.11 to 2.13 (which is our plan) you
have to do an offline reindex in the middle of the upgrade before
starting on 2.13. You can only do online reindexes if you step through
2.12 as well.

The default for offline_reindex is to not perform one because it is
slow, but we know we need one here so flip the boolean value while we
sort out the upgrade.

Change-Id: I4e43308defee12d17893ca8ff541aa7c1b49d127