Upgrade review-dev to Gerrit 2.13.7
Update the war installed on review-dev so that puppet will upgrade us
from Gerrit 2.11 to 2.13.

Note that this upgrade requires you either upgrade to 2.12 first or do
an offline reindex prior to starting 2.13 so the previous change has set
offline_reindex to true. You must also init the site which will update
the database schema. With the previous change both of these items should
be autmatically performed by the puppetry, automating this upgrade.

Change-Id: I96e637722d3289142b92b37ecb45fd5bacba24ef