Correct the ca-cert for mqtt on subunit worker
This patch fixes a mistake on my part, previously the location of
the ca certs on archlinux, but on ubuntu this isn't where they
live. So the mqtt libs can't properly encrypt the traffic and
things are getting stuck in a loop and not working. This commit fixes
this by writing the cert from instantssl/comodo to disk and then
pointing the worker at that. This way it should always work.

Change-Id: I21b1a64b457545115ff862e3c3388c5892c5497b