WIP: Use ansible for zuulv3.openstack.org
This could be an example of how we could bootstrap a server with
puppet (since playbooks don't exist). Then have ansible manage the
server.

The idea is we'd still rely on server.pp from puppet, which sets up
users / iptables / mta /etc.  Once the server is online, we'd then
configuration the remaining softrware using ansible.  This process
will allow us to slowly migrate our server to ansible.

Secrets will still be managed in hieradata, this means we'll have to
read the contents of the yaml files already on the remote server when
populating things like SSH keys.

[1] http://git.openstack.org/cgit/openstack/ansible-role-zuul

Change-Id: I9db17604538fec2f55475ee863514d341a52d72e
Signed-off-by: Paul Belanger <pabelanger@redhat.com>