WIP: Use ansible for zuulv3.openstack.org
This could be an example of how we could bootstrap a server with
puppet (since playbooks don't exist). Then have ansible manage the
server.

In this example, we are missing things like SSH keys for zuul users,
logging configuration and zuul configuration. This can be added, but
what a first glimpse looks like.

We'd also need to get the openstack.zuul role onto puppetmaster.o.o.
But, this does work today with a 2 stage launch.  Use launch-node.py
to create server, enroll into puppetmaster and have ansible run on the
node.  Kinda like we do with infracloud today.

[1] http://git.openstack.org/cgit/openstack/ansible-role-zuul

Change-Id: I9db17604538fec2f55475ee863514d341a52d72e
Signed-off-by: Paul Belanger <pabelanger@redhat.com>