Add firehose documentation for notifications
This commit adds docs on event notifications as an example use case.
This is a common thing that people want to do with firehose events, so
it's good to document how to do it. This commit doesn't go into detail
on how to use/configure mqttwarn (since it can be quite involved) but
points people at the docs.

Change-Id: Ic5fff993e5f268e2dbd82062b4fb95483d69abf2