Add a trackingid for Launchpad bug headers
Add a Launchpad trackingid in Gerrit for our standard
{Closes,Partial,Related}-Bug headers so that the tr/bug search
operator can be used to find them easily. This will need not only a
Gerrit restart (because it's changing the gerrit.config file), but
also a reindex (online reindex works fine for this).

The resulting configuration has already been tested on
review-dev.openstack.org and works as intended.

Change-Id: I113cb139ff5df1d6ec05f060dd3330ab44458a89
Depends-On: I99e16246e81b7ebc50e96054a2e3790aed0089bc