Add missing cert checks
Based on an audit of certs I currently have record of ordering,
update the ssldomains list for our cert check to include everything
I know (and can confirm) is in production. Drop security since it's
rolled into the SubjectAltName set for the static.openstack.org cert
now. Also remove groups-dev from the list as it's the only one using
a snakeoil cert and we don't normally bother to check for expiration
on those. Keep www on the list even though Infra doesn't manage that
one, because its unanticipated expiration has impacted us in the
past and having a bit of extra warning there can't hurt.

Change-Id: I4a51d0cd15533a39d23e09735c9fda34398e957f