Adding tripleo check jobs into logstash
Although is not common to have a check gate into this, the tripleo
project only run jobs in other projects with non-voting status, which
means, sometimes a job that fails under tripleo is actually a bug that
might block the tripleo ci. Having this patch applied, will allow
tripleo jobs to be indexed in openstack-health, and so allow the
developers who can subscribe to the rss feed to be more pro-active when
some job fail, before block the whole ci.

Change-Id: I08dd5239ca3597334669ef6f61b0564a7d997837