Add vhost on static.o.o for service-types.o.o
We want to publish json information from the service-types-authority to
service-types.openstack.org. We need a vhost for that, and also a
directory for the post job to copy data to.

This also puts the OpenStack CSS files in at the mod_autoindex level.
That's not ideal, but there's not a great way to inject them into the
generated HEADER.html file in any case.

Change-Id: Ia53a6e508e4b6d79fab08b4c51dc3ad97a929502