Update the keystone files in logstash worker config
Over time the keystone log file location has evolved in part due to
keystone dropping support for eventlet and relying on wsgi servers to
host keystone. We have gone through the following iterations of log
locations:

  logs/screen-key.txt # Eventlet based log file
  logs/apache/keystone.txt # Apache + wsgi < pike location
  logs/screen-keystone.txt # Apache + uwsgi + system >= pike location

Since eventlet is completely gone at this point we drop all support the
the screen-key.txt log name. We add in support for screen-keystone.txt
which was missing and update the comments around apache/keystone.txt to
indicate it is < pike (so will only be present on older branches and in
grenade jobs).

Change-Id: I8b04a8ee6c6ab45ffac47119df307a86cd3c2c23