Include retired projects/deliverables in elections
If a deliverable or entire project is retired within the qualifying
period for an election, those contributions still count toward
defining the electorate. Merge them into the list of official
projects under those circumstances, but check for any possible
duplicates in case of partial deliverable retirement for a
multi-repo deliverable.

Change-Id: If626373cdbc6ca9ad7d016b2e18b4d82a4ceaa23
Depends-On: I7ca3c9294d325085ce9139f3a86b55e5b31c79c7