Add proxy-initial-not-pooled to proxy
As described in the inline comment, we are seeing infrequent 502
errors in the logs that do seem to match what is described as a
race-condition in the Apache documentation.

As suggested there, enable proxy-initial-not-pooled in an attempt to
avoid these.

It says I"t has to be kept in mind that setting this variable
downgrades performance, especially with HTTP/1.0 clients"
... downgraded performance would be preferable to 502 errors leading
to long-running CI jobs failing randomly; so seems like a trade-off
worth trying.

Change-Id: I94a73e0fc07ca5b8903f4e74ad7c224509cea323