Make journal logs persistent on disk
By default journald's storage scheme is "auto" which means if
/var/log/journal exists log to it on disk otherwise use the ring buffer.
Problem is we don't have a /var/log/journal on our nodes.

Rather than just create /var/log/journal we instead set the scheme to
persistent so that journald can configure /var/log/journal for us
automatically with the correct ownership and permissions and acls.

Change-Id: I68116a04d8f7a8e681242f6e13dd8b412fcae63a