Add packages.erlang-solutions.com reverse proxy
This patch adds a reverse proxy for rpm/deb packages hosted on
packages.erlang-solutions.com. This is the trusted source for erlang
packages that are approved for use with RabbitMQ.

The erlang-solutions mirror is located in Eastern Europe and causes
significant delays in OpenStack-Ansible gate jobs that must install
erlang dependencies for RabbitMQ. This change should allow gate jobs
to reduce timeouts.

Change-Id: Ife4b2e5e8ee8f356bd74e8aa980fd2cc3e9b2724