Update infracloud-chocolate public ssl cert
Our old self signed ssl cert will expire soon. Rather than need to maintain
a CA for infracloud or sort out how we want to sign things lets just go
ahead and use our snakeoil cert on the controller. The two big drawbacks
to this approach are the long length of time this cert is valid for and
it forces us to use a single controller host per cloud.

Note that private hiera will have to be updated with the private key
that corresponds to this (is the private portion of the snake oil cert
on controller00.chocolate).

Change-Id: Iad87881f47354e013c65866dd0138b6bc1cc7f70