Trust infracloud directly using public cert
Rather than needing to manage /usr/local/share/ca-certificates and keep
the global cert list up to date with update-ca-certificates just
directly trust the appropriate cert when talking to infracloud. This is
how we have done it on nodepool and it is a very simple method for
managing this setup.

Change-Id: I702e30de81f1ef3211caa113616fef0be51f4821