Switch review.o.o to cgit links
It was noticed that review.o.o doesn't have the "(cgit)" links next to
commit/parent hashes in the commit box.

It looks like the inbuilt gitweb integration is returning 404 for some
reason after the upgrade.  One option is to use the cgit integration
instead.  This seems to make some sense as cgit is behind a
load-balancer and may perform better than local cgi calls?

Note review-dev uses cgit like this, but that may be just
unintentional.

Change-Id: I86d8e724b22ee349f58540864888926d766daf1d