Set ZooKeeper purge_interval for on nodepool.o.o
Set ZooKeeper's autopurge.purgeInterval to 6 hours for
nodepool.openstack.org to limit the number of ZK snapshots kept.
These will quickly fill the disk in production if this is not
enabled.

The purge_interval class parameter determines the frequency in hours
to look for and purge old snapshots, defaulting to 0 (disabled). The
number of retained snapshots can be separately controlled through
snap_retain_count but already defaults to a sane value of 3 (the
minimum supported for autopurge.snapRetainCount).

Change-Id: If84094641add539a815718ae604729fe6f0545cd