Create more inodes on ephemeral partition fs
When we create the ext4 filesystem on the ephemeral device presented to
rax nodes do so with roughly twice the number of inodes as the default
ext4 bytes to inodes ratio.

We do this because we have found that we have in multiple cases needed
more inodes and that is difficult to deal with after the fact. This may
create more inodes than necessary in other cases but the cost is
relatively low (just more disk used).

Change-Id: Ie5414212f4505e98ba3ee8144ac90fcd05aa5b67