Add new backup host to puppet
Add definition for a new backup host.  We'll use ci.openstack.org to
make it clear it's not general-purpose, and more mirror-like name in
the form

 backup01.region.provider.ci.opentack.org

We will rsync the backup data from the old host to this new host
before cleaning up the old one (although it's on cinder volumes that
could be re-attached, it seems like a good opportunity to verify
reading all the data anyway).

Change-Id: I7d0c96ef6f1c68d977da70c047065b10f975b4f2