Add new backup host to puppet
Add definition for a new backup host.  We'll use ci.openstack.org to
make it clear it's not general-purpose, and more mirror-like name in
the form

 backup01.region.provider.ci.opentack.org

We have decided to start with fresh backups, so this is using a fresh
3TiB LVM of cinder volumes.  After removing the old server, we will
attach the old volumes to this for a period in case we need it.

Change-Id: I7d0c96ef6f1c68d977da70c047065b10f975b4f2