Remove ci-backup-rs-ord.openstack.org
Migrate backups to new backup01.ord.rax.ci.openstack.org

We decided to start fresh backups on the new server, so this is ready
to go.  I have performed an initial backup on each server so it has
accepted the host key of the new server and been tested (I also fixed
up review-dev.o.o, which was rebuilt but keys not updated ... todo:
add this to puppet, but since it changes so infrequently not high
priority).

Change-Id: I0872f9fcf4a334d32f632b3cb04801deefab4fd1