Remove zl's from jenkins-logstash-client config
The zuul launcers are gone so we can't have the jenkins logstash client
listen to them for 0mq events anymore. This is fine we submit the jobs
directly from Zuulv3 now and just need this to act as a geard.

Future cleanup should replace this service with a standalone geard
instead.

Change-Id: I132f065d570054081a05b27a430cfd3a5a2ad461