Disable crm114 on the logstash worker nodes
The logstash worker nodes are running out of disk because crm114 has
enough unique files to process that at 13MB per file we are still
filling up many gigabytes of data and filling out disks.

We will need to work on collapsing files (perhaps not use their full
path and just use their basename?) and work to eliminate log files that
are unique like those that have the hostname in them (as this means we
create a new file for every job run).

Change-Id: Ie729a3d2628f0ded65574c267e31c14f6b795b29