Allow logs.openstack.org to be used with storyboard-dev
The js-draft build for storyboard-dev uploads things to logs.o.o
instead of docs-draft.o.o now. Currently this results in a broken
experience because logs.o.o is not in cors_allowed_origins for
storyboard-dev.

This commit adds logs.o.o to both cors_allowed_origins and
valid_oauth_clients in order to make the draft UI functional again.

Change-Id: I0b0b644145389ef885d40dd00a9e2a0757f08cdd