Timeout logstash es pushes after 300s
By default the elasticsearch output plugin for logstash has no timeout.
I think this is the cause of logstash daemons hanging when we have
elasticsearch node outages. Goal here is to have the timeout force the
request to be retried again (ideally against a different server entirely
but even if it is simply retried to the same server we should operate
properly again once the server that went down is back up).

Change-Id: I0511f32f8ee3f0c40987cab3e2528a85595785b2