Better checking for tags when cloning puppet modules
One of the puppet modules changed tags from "2.0" to "v2.1" which I
missed.  The testing was rather unhelpful ... when "2.1" didn't exist
it went ahead and left it checked out on master and then ran the test
suite against that.  So it was quite confusing as to what the root
cause was.

Do a more thorough check of the wanted tag; if it's not there at
checkout time, abort.

Change-Id: Ic358857c56c0b5a95f1052c21bfb89e17a9816cc