Reduce cacti updates cron spam
The current process to udpate cacti hosts is fairly chatty. It basically
goes in and just tries to update every host there and if they exist that
generates a bunch of output which is then emailed to infra roots. This
information is potentially useful for debugging so keep it around in
local cacti host logs. These logs will then be rotated with a week of
retention.

This should help make our inboxes happier.

Change-Id: Ib03ef7b22083a2a2454715bd5229313b19b84ae9