Setup bash history to contain standard timestamps (RFC3339)
The bash history is an unintended form of documentation and also a way
to see what happened but by default it doesn't tell you when.

Before this change:
  $ history
      1  uname -a
      2  sudo reboot

After this change:
  $ history
      1  2017-12-20T23:50:28+0000 uname -a
      2  2017-12-20T23:50:35+0000 sudo reboot

Note that any entries in the bash history before this change will be
defaulted to when the change takes place.

Change-Id: I4443f00ab050891a16e545315ee88ae24893ac5d