Set the vhost_name for zuul01 to zuul.openstack.org
The vhost_name defaults to the fqdn which does not match
zuul.openstack.org and therefore does not match the SSL certificate.
The reason that it currently works is because it is the only Apache
vhost so it grabs everything.

The side effect is that the backup script is attempting to retrieve
the status.json over https://zuul01.o.o and it fails certificate
validation.

Change-Id: Ied4a97f875177098bd4663c1e7d880004da5a656