Allow multiple declarations of ARM64 of HWE kernel package
The openafs puppet module also declares a linux-generic-hwe-16.04
package for arm64 as it is required for those modules.  This is to
cover the non-afs server case, where the later kernel still works
better anyway.  Switch to ensure_packages, which handles if it is
already declared, so everyone can live together in peace.

Change-Id: I72c9423956b7739695a04a5de27f5d89c67240d0