Make an entry for zuul-web-test
Debugging the issue with zuul-web and apache from the recent javascript
change is weird. Make a test server so we can run the puppet as close to
production as possible, but only run zuul-web and not a scheduler.

When we're done investigating the issue, we can delete the server again
and revert this patch.

Change-Id: I04ca76872da7aebeb5f900c73b28b3acd30a9725