Sync centos from limestone networks mirror
The kernel.org mirror's rsync server is overwhelmed with connections and
we are not getting updates to centos in a timely manner. Update the
upstream source mirror to use limestone network's mirror which should be
fairly close to our mirror infrastructure and we have friends there now
that can hopefully sort out problems if they arise.

Change-Id: Icb70b4033abee3a7d1893d7a1b53216940362f63