Handle job results without urls
The existing regex assumes "- name url : details" will be turned in to
<li>name<a href=url>url : details</li> by gerrit's processing. However,
for jobs that have a POST_FAILURE that winds up without a valid log_url,
the url doesn't get turned into an <a> so the regex doesn't do what it's
supposed to.

The zuul change below changes the return value to list "None" for url
when there is no url. Add a regex that looks for this and otherwise
updates the output html to look how people expect.

Change-Id: If068e5c17a509736eb304a8bb853f9c658c2a892
Depends-On: https://review.openstack.org/554059