Update Gerrit project renaming for Zuul v3
With the switch to Zuul v3, we need to resolve some configuration
catch-22s where project names and related in-repository job
definitions can't happen without a complex multi-stage removal and
reintroduction process to get it through speculative testing
successfully. For now, just punt and use monolithic changes
bypassing CI in code review. As an up side, the Ansible automation
of this process coupled with Zuul v3's increased resilience to
on-the-fly configuration changes means we can skip stopping/starting
it now and significantly simplify the process.

Since we're here, correct the section heading level for
"Force-Merging a Change" in the sysadmin document.

Change-Id: I335c23abd0b5706f43bbea2dd8cfffa4280dd5db