WIP: Rewrite launch-node.py in Ansible playbooks/roles
launch-node is more or less a wrapper for shade and Ansible.
It would be awesome to use Ansible and OpenStack modules which happen
to use Shade underneath.

This patch introduces the necessary default variables in group_vars/all
and host_vars for (the first) host using these playbooks.
Keeping the host variables up to date would allow us to easily (or even
automatically) redeploy hosts based on their configuration for cloud,
region, flavor, fqdn, etc.

Prior to this, launch-env was pinned to Ansible 2.1 and Shade 1.11.
This is being developed against Ansible 2.5 and Shade 1.27.x to future
proof this work.

Change-Id: I4ba636d9873a802b0df873037a37abcd0b588896