Revert "centos7: don't install EPEL"
This reverts commit 37c26280bfba775acfd1455be895ef6488efbbb5.

This was originally removed so that EPEL would not be installed in
our nodepool images.  We no longer run this script when creating
those images, so it can be safely restored.  It should be restored
because without this, we are unable to spin up a cgit server
(which requires epel).

Change-Id: I93e6487ab9a2eae5f927abf43aea7e3e19c5ac9e