Pass whitelabel tenant to zuul::web
In v2 the web content was tied to the scheduler, so the web config was
in ::zuul. In v3, zuul-web is separate and needs vhost config to be able
to flow from the ::zuul::web manifest.

While we're at it, remove a few other settings that are no longer
needed, and move the ssl cert settings to the right place.

Depends-On: https://review.openstack.org/556869
Change-Id: I355254e68e41aa7d9af9417bce55701285737f75