Update connection construction in launch_node
The dns.py file uses openstack.connect to make the Connection but
launch_node.py was still using shade.OpenStackCloud, so when the
connection was passed to dns.py it was trying to use an SDK property but
getting a Shade object.

This is because while sdk has been updated with all of the shade objects,
we haven't updated shade yet to provide the sdk version of the object, so
shade objects from sdk have things shade objects from shade don't yet have.

Update launch_node.py to use the same Connection construction that
dns.py does.

Change-Id: I1c6bfe54f94effe0e592280ba179f61a6d983e7a