Minor fork puppet-solr
Upstream seems abandonded; Xenial support was only proposed in January
with [1] (suggesting nobody is really using this on modern systems)
and remains unmerged.  Correspondence hasn't been responded to as yet.

As it is MIT licensed, it's not suitable for importing as is.

Thus I have made a minor fork and added the few changes required for
basic Xenial support.  This is a requirement of our askbot
installation.  It has passed a basic apply rspec test in [2].

I had a look at other modules, but the other top downloaded puppet
module *is* abanonded, according to it's readme.

[1] https://github.com/vamsee/puppet-solr/pull/33
[2] https://review.openstack.org/558991

Change-Id: Iee7332879fa036890be20ba9d8618e61a47fabf9